fastapi_project/requirements.txt
fastapi
uvicorn
vanna
qdrant-client
psycopg2-binary
python-dotenv
pandas
openai